make sure the CI instances get only placed into the admin crestash policy
The TF juggling required here is nasty.

 var.enviromnet =~ s/[^,]//
    admin,stage,prod => ,,

Then we concatenate with a regular expression again, the single IAM role
returned by CI:

  var.ci_role =~ s/$/,,/

So finally, we have a string that looks like:
  ci-role-arn,,

Then, we can index in there with element(var.ci_role, count.index)

And we'll get the correct ARN in the first VPC, them an empty value
in the other ones.

It's fugly, I know, and TF 0.7 will make all that juggling unnecessary