Manage user_management config file
First see nubisproject/nubis-deploy/138, this patch is pretty much the
same as that PR except we are splitting up the enable functionality, we
can enable user_management for consul and IAM seperately. For now I'm
setting IAM to default to disable.

This commit will generate a config file for the user_management lambda
function. It will generate a config file and upload it to kms using
credstash