Ability to accept multiple groups for user_management (#171)
User management should be able to dish out multiple groups, this patch
fixes that issue by fixing the configs to be able to take multiple
groups. We accept the list as a comma seperated list but when terraform
generates the template it translates "," to "|" so that it doesn't look
weird for end users.