Trim down admin users's IAM policy
Basically, since we need to sts:AssumeRole before we can do anything of interest,
and that the admin user isn't an admin anymore, all we need is sufficient permissions
to setup/manage our MFA device, and assume one of the admin or read-only roles.

So it looks like:
 - iam:ListVirtualMFADevices
 - iam:CreateVirtualMFADevice (only ours)
 - iam:EnableMFADevice (only ours)
 - iam:ResyncMFADevice (only ours)
 - sts:AssumeRole
   - role/nubis/admin/$username
   - role/nubis/readonly

And nothing else, leaving AWS to default Deny everything else for us.

Of course, once MFA'ed into our admin role, we can still do whatever we want.